% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalEncoders.R
\name{get_variable_summary}
\alias{get_variable_summary}
\title{Get Variable Summary}
\usage{
get_variable_summary(dataset)
}
\arguments{
\item{dataset}{A \code{data.frame} containing the dataset to be processed.}
}
\value{
A data frame with two columns: 'Qualitatives' and 'Quantitatives', containing the names of the qualitative and quantitative variables, respectively. If one type has fewer variables, the remaining rows are filled with NA.
}
\description{
This function returns a summary of the qualitative and quantitative variables in the dataset.
}
