% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalEncoders.R
\docType{class}
\name{CategoricalVerifier}
\alias{CategoricalVerifier}
\title{CategoricalVerifier Class}
\description{
A class to handle encoding of categorical variables in a dataset.

This class provides various encoding methods such as label encoding, one-hot encoding,
frequency encoding, and binary encoding. It can dynamically apply these encodings based
on user-defined preferences or defaults.
}
\examples{
# Load Titanic dataset

library(ggplot2)

data <- as.data.frame(Titanic)
data <- data[rep(seq_len(nrow(data)), data$Freq), ]
data$Freq <- NULL
data$Survived <- factor(data$Survived, levels = c("No", "Yes"))
data$Class <- factor(data$Class)
data$Sex <- factor(data$Sex)
data$Age <- factor(data$Age)

library(LogRegSisem2)

# Choose any of the categories available to encode: 
# label = label encoding (Default), 
# one_hot = One-Hot Encoding, 
# frequency = frequency encoding, 
# binary = binary encoding

# Create a Hash table of encoding for variables we want a different encoding than 
# the default one (Label encoding)
encoding_dict = list(Sex = "frequency", Age = "binary")

# Create the categorical verifier
verifier <- CategoricalVerifier$new(data, encoding_dict = encoding_dict)

# Apply the encoding
verifier$apply_encoding()


# Note:
# If you don't specify any encoding method, it will use default encoding (Label Encoding).

# Get the processed dataset
processed_data <- verifier$get_dataset()
print(processed_data)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset}}{The dataset to be processed.}

\item{\code{categorical_vars}}{Names of categorical variables in the dataset.}

\item{\code{encoded_vars}}{A list of encoded variables.}

\item{\code{encoding_dict}}{A dictionary specifying encoding methods for each variable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategoricalVerifier-new}{\code{CategoricalVerifier$new()}}
\item \href{#method-CategoricalVerifier-verify_categorical_vars}{\code{CategoricalVerifier$verify_categorical_vars()}}
\item \href{#method-CategoricalVerifier-label_encoding}{\code{CategoricalVerifier$label_encoding()}}
\item \href{#method-CategoricalVerifier-one_hot_encoding}{\code{CategoricalVerifier$one_hot_encoding()}}
\item \href{#method-CategoricalVerifier-frequency_encoding}{\code{CategoricalVerifier$frequency_encoding()}}
\item \href{#method-CategoricalVerifier-binary_encoding}{\code{CategoricalVerifier$binary_encoding()}}
\item \href{#method-CategoricalVerifier-apply_encoding}{\code{CategoricalVerifier$apply_encoding()}}
\item \href{#method-CategoricalVerifier-get_dataset}{\code{CategoricalVerifier$get_dataset()}}
\item \href{#method-CategoricalVerifier-clone}{\code{CategoricalVerifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-new"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-new}{}}}
\subsection{Method \code{new()}}{
Initialize the CategoricalVerifier class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$new(dataset, encoding_dict = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{A \code{data.frame} containing the dataset to be processed.}

\item{\code{encoding_dict}}{A named list specifying encoding methods for categorical variables.
Defaults to NULL, which uses label encoding for all variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-verify_categorical_vars"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-verify_categorical_vars}{}}}
\subsection{Method \code{verify_categorical_vars()}}{
Identify categorical variables in the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$verify_categorical_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector containing the names of categorical variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-label_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-label_encoding}{}}}
\subsection{Method \code{label_encoding()}}{
Apply label encoding to a variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$label_encoding(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{A character string representing the variable name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-one_hot_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-one_hot_encoding}{}}}
\subsection{Method \code{one_hot_encoding()}}{
Apply one-hot encoding to a variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$one_hot_encoding(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{A character string representing the variable name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-frequency_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-frequency_encoding}{}}}
\subsection{Method \code{frequency_encoding()}}{
Apply frequency encoding to a variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$frequency_encoding(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{A character string representing the variable name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-binary_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-binary_encoding}{}}}
\subsection{Method \code{binary_encoding()}}{
Apply binary encoding to a variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$binary_encoding(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{A character string representing the variable name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-apply_encoding"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-apply_encoding}{}}}
\subsection{Method \code{apply_encoding()}}{
Apply specified encodings to all categorical variables in the dataset.
If \code{encoding_dict} is NULL, label encoding is applied to all variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$apply_encoding()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing the dataset with encoded variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Retrieve the processed dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing the processed dataset.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalVerifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalVerifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalVerifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
