% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic Regression Model with Gradient Descent Optimization}
\description{
This R6 class implements a logistic regression model with gradient descent optimization.
}
\section{Fields}{

\describe{
\item{\code{learning_rate}}{The learning rate for gradient descent. Default is 0.01.}

\item{\code{iterations}}{The number of iterations for gradient descent. Default is 1000.}

\item{\code{X}}{The feature matrix with an added intercept term.}

\item{\code{y}}{The response vector.}

\item{\code{beta}}{The coefficients of the logistic regression model.}

\item{\code{n_features}}{The number of features in the model.}

\item{\code{ll}}{The log-likelihood of the model.}
}}

\examples{
# Generate synthetic dataset similar to your previous example
# But now we'll use iris dataset for a more standard comparison
data(iris)

# Convert to binary classification (setosa vs. others)
iris_binary <- iris[iris$Species != "versicolor", ]
iris_binary$Species <- ifelse(iris_binary$Species == "setosa", 1, 0)

# Prepare features and target
X <- as.matrix(iris_binary[, c("Sepal.Length", "Sepal.Width")])
y <- iris_binary$Species

# Split data
train_indices <- createDataPartition(y, p = 0.8, list = FALSE)
X_train <- X[train_indices, ]
y_train <- y[train_indices]
X_test <- X[-train_indices, ]
y_test <- y[-train_indices]

# Method 1: Custom R6 Logistic Regression
reglog_model <- LogisticRegression$new(learning_rate = 0.01, iterations = 1000)
reglog_start_time <- Sys.time()
reglog_model$fit(X_train, y_train)
reglog_pred_proba <- reglog_model$predict_proba(X_test)
reglog_pred_class <- reglog_model$predict(X_test)
reglog_end_time <- Sys.time()

# Calculate R6 Model Performance
reglog_accuracy <- mean(reglog_pred_class == y_test)
reglog_log_likelihood <- reglog_model$get_log_likelihood()
reglog_coefficients <- reglog_model$get_coefficients()
reglog_computation_time <- as.numeric(reglog_end_time - reglog_start_time)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogisticRegression-new}{\code{LogisticRegression$new()}}
\item \href{#method-LogisticRegression-fit}{\code{LogisticRegression$fit()}}
\item \href{#method-LogisticRegression-predict_proba}{\code{LogisticRegression$predict_proba()}}
\item \href{#method-LogisticRegression-predict}{\code{LogisticRegression$predict()}}
\item \href{#method-LogisticRegression-log_likelihood}{\code{LogisticRegression$log_likelihood()}}
\item \href{#method-LogisticRegression-get_coefficients}{\code{LogisticRegression$get_coefficients()}}
\item \href{#method-LogisticRegression-get_log_likelihood}{\code{LogisticRegression$get_log_likelihood()}}
\item \href{#method-LogisticRegression-clone}{\code{LogisticRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$new(learning_rate = 0.01, iterations = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learning_rate}}{The learning rate for gradient descent. Default is 0.01.}

\item{\code{iterations}}{The number of iterations for gradient descent. Default is 1000.
Train the model using gradient descent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-fit"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The feature matrix.}

\item{\code{y}}{The response vector.
Predict probabilities}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$predict_proba(new_X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_X}}{The new feature matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predicted probabilities.
Predict classes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-predict"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$predict(new_X, threshold = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_X}}{The new feature matrix.}

\item{\code{threshold}}{The threshold for class prediction. Default is 0.5.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predicted classes.
Log-likelihood calculation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-log_likelihood}{}}}
\subsection{Method \code{log_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$log_likelihood()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The log-likelihood of the model.
Getter for coefficients
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-get_coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-get_coefficients}{}}}
\subsection{Method \code{get_coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$get_coefficients()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The coefficients of the logistic regression model.
Getter for log-likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-get_log_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-get_log_likelihood}{}}}
\subsection{Method \code{get_log_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$get_log_likelihood()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The log-likelihood of the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogisticRegression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogisticRegression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
